requirements.txt
numpy==1.26.4
pandas==2.1.4
pyarrow==14.0.2

streamlit==1.32.2
plotly==5.20.0
matplotlib==3.8.2

openai==1.35.13
python-dotenv==1.0.1
requests==2.31.0

reportlab==4.0.4

# watchdog compatible con macOS Catalina (sin fallas en fsevents)
watchdog==3.0.0

pip>=23.2.1
setuptools>=68.2.2
wheel>=0.41.2
tabulate
